Estimate emission reductions under AMS-I.E

Chains the AMS-I.E equation helpers to translate monitoring data into emission reductions.

estimate_emission_reductions_ams_ie(biomass_data,
  consumption_col = "biomass_consumption_tonnes",
  fraction_col = "non_renewable_fraction", fraction = 1, ncv,
  group_cols = NULL, emission_factor, project_energy_col = NULL,
  project_emission_factor = 0)

Arguments

biomass_data

Tibble containing monitoring observations.

consumption_col

Column storing biomass consumption.

fraction_col

Column storing the non-renewable fraction.

fraction

Scalar non-renewable fraction used when fraction_col is NULL.

ncv

Net calorific value in MJ per unit biomass.

group_cols

Grouping columns identifying entities.

emission_factor

Baseline emission factor in tCO2e/MJ.

project_energy_col

Optional column storing project fossil energy in MJ.

project_emission_factor

Project emission factor in tCO2e/MJ.

Value

A tibble containing baseline energy, baseline emissions, project emissions, and emission reductions.

Examples

monitoring <- tibble::tibble(
  site_id = c("A", "A", "B"),
  biomass_consumption_tonnes = c(10, 6, 12),
  non_renewable_fraction = c(0.8, 0.75, 0.9),
  project_energy_mj = c(100, 120, 140)
)
estimate_emission_reductions_ams_ie(
  monitoring,
  group_cols = "site_id",
  ncv = 15,
  emission_factor = 0.0001,
  project_energy_col = "project_energy_mj",
  project_emission_factor = 0.00009
)
#> # A tibble: 2 × 7
#>   site_id non_renewable_biomass_tonnes baseline_energy_mj baseline_emissions_t…¹
#>   <chr>                          <dbl>              <dbl>                  <dbl>
#> 1 A                               12.5               188.                 0.0188
#> 2 B                               10.8               162                  0.0162
#> # ℹ abbreviated name: ¹​baseline_emissions_tco2e
#> # ℹ 3 more variables: project_energy_mj <dbl>, project_emissions_tco2e <dbl>,
#> #   emission_reductions_tco2e <dbl>